fix side bar buttons icons
