+++ src/renderer/components/Icons.tsx
@@ -1,7 +1,17 @@
 import erwt from '@assets/icons/erwt.png';
-import LogoIcon from '@assets/icons/logo.svg';
+import logo from '@assets/icons/logo.svg';
+import search from '@assets/icons/search.svg';
+import notebook from '@assets/icons/notebook.svg';
+import mathPanel from '@assets/icons/mathPanel.svg';
+import archive from '@assets/icons/archive.svg';
+import settings from '@assets/icons/settings.svg';
 
 export const icons = {
   erwt,
-  LogoIcon
+  logo,
+  search,
+  notebook,
+  mathPanel,
+  archive,
+  settings,
 };
